Module pabi::chess::attacks

Mappings of occupied squares to the attacked squares for each piece. The mappings are pre-calculated where possible to provide an efficient way of generating moves.

Structs

AttackInfo

Constants

BISHOP_ATTACKS

BISHOP_ATTACKS_COUNT

BISHOP_ATTACK_OFFSETS

BISHOP_RAYS

BISHOP_RELEVANT_OCCUPANCIES

BLACK_LONG_CASTLE_KING_WALK

BLACK_LONG_CASTLE_ROOK_WALK

BLACK_PAWN_ATTACKS

BLACK_SHORT_CASTLE_KING_WALK

BLACK_SHORT_CASTLE_ROOK_WALK

KING_ATTACKS

KNIGHT_ATTACKS

RAYS

ROOK_ATTACKS

ROOK_ATTACKS_COUNT

ROOK_ATTACK_OFFSETS

ROOK_RAYS

ROOK_RELEVANT_OCCUPANCIES

WHITE_LONG_CASTLE_KING_WALK

WHITE_LONG_CASTLE_ROOK_WALK

WHITE_PAWN_ATTACKS

WHITE_SHORT_CASTLE_KING_WALK

WHITE_SHORT_CASTLE_ROOK_WALK

Functions

bishop_attacks

bishop_ray

king_attacks

knight_attacks

pawn_attacks

pext

queen_attacks

ray

rook_attacks

rook_ray