Module pabi::chess::bitboard

Bitboard-based representation for crate::chess::position::Position. Bitboards utilize the fact that modern processors operate on 64 bit integers, and the bit operations can be performed simultaneously. This results in very efficient calculation of possible attack vectors and other meaningful features that are required to calculate possible moves and evaluate position. The disadvantage is complexity that comes with bitboard implementation and inefficiency of some operations like “get piece type on given square” (efficiently handled by Square-centric board implementations that can be used together bitboard-based approach to compensate its shortcomings).

Structs

Bitboard

Represents a set of squares and provides common operations (e.g. AND, OR, XOR) over these sets. Each bit corresponds to one of 64 squares of the chess board.

BitboardIterator

Iterates over set squares in a given Bitboard from least significant 1 bits (LS1B) to most significant 1 bits (MS1B) through implementing bitscan forward operation.

Board

Piece-centric implementation of the chess board. This is the “back-end” of the chess engine, an efficient board representation is crucial for performance. An alternative implementation would be Square-Piece table but both have different trade-offs and scenarios where they are efficient. It is likely that the best overall performance can be achieved by keeping both to complement each other.

Pieces

Piece-centric representation of all material owned by one player. Uses Bitboard to store a set of squares occupied by each piece. The main user is crate::chess::position::Position, Bitboard is not very useful on its own.

Constants

LINE_SEPARATOR

SQUARE_SEPARATOR