Struct pabi::chess::bitboard::Bitboard

pub struct Bitboard {
    bits: u64,
}

Represents a set of squares and provides common operations (e.g. AND, OR, XOR) over these sets. Each bit corresponds to one of 64 squares of the chess board.

Mirroring Square semantics, the least significant bit corresponds to A1, and the most significant bit - to H8.

Bitboard is a thin wrapper around u64.

Fields
bits: u64
Implementations

impl Bitboard

pub(super) const fn from_bits(bits: u64) -> Self

Constructs Bitboard from pre-calculated bits.

pub const fn empty() -> Self

Constructs a bitboard representing empty set of squares.

pub const fn full() -> Self

Constructs a bitboard representing the universal set, it contains all squares by setting all bits to binary one.

pub const fn bits(self) -> u64

Returns raw bits.

pub(super) fn from_squares(squares: &[Square]) -> Self

pub(super) fn extend(&mut self, square: Square)

Adds given square to the set.

pub(super) fn clear(&mut self, square: Square)

Adds given square to the set.

pub(super) fn contains(self, square: Square) -> bool

Returns true if this bitboard contains given square.

pub(super) fn as_square(self) -> Square

pub(super) fn count(self) -> u32

pub(super) fn is_empty(self) -> bool

pub(super) fn has_any(self) -> bool

pub(super) fn shift(self, direction: Direction) -> Self

pub(super) fn iter(self) -> BitboardIterator

An efficient way to iterate over the set squares.