Type Alias pabi::chess::core::MoveList

pub type MoveList = ArrayVec<Move, { MAX_MOVES }>;

Moves are stored on stack to avoid memory allocations and improve performance. This is important for performance reasons and also prevents unnecessary copying that would occur if the moves would be stored in std::Vec with unknown capacity.

Aliased Type

struct MoveList {
    xs: [MaybeUninit<Move>; 256],
    len: u32,
}

Fields
xs: [MaybeUninit<Move>; 256]len: u32