Module pabi::chess::position

Provides fully-specified Chess Position implementation: stores information about the board and tracks the state of castling, 50-move rule draw, etc.

The core of Move Generator and move making is also implemented here as a way to produce ways of mutating Position.

Structs

Position

State of the chess game: board, half-move counters and castling rights, etc. It has 1:1 relationship with Forsyth-Edwards Notation (FEN).

Functions

generate_bishop_moves

generate_castle_moves

generate_king_moves

generate_knight_moves

generate_pawn_moves

generate_rook_moves

perft

Perft (performance testing) is a technique for checking correctness of move generation by traversing the tree of possible positions (nodes) and calculating all the leaf nodes at certain depth.

validate