Module pabi::chess::attacks

Mappings of occupied squares to the attacked squares for each piece. The mappings are pre-calculated where possible to provide an efficient way of generating moves.

The implementation uses BMI2 (if available) for performance (reference), specifically the PEXT instruction for PEXT Bitboards.

Structs

AttackInfo 🔒

Constants

BLACK_LONG_CASTLE_KING_WALK 🔒

BLACK_LONG_CASTLE_ROOK_WALK 🔒

BLACK_SHORT_CASTLE_KING_WALK 🔒

BLACK_SHORT_CASTLE_ROOK_WALK 🔒

WHITE_LONG_CASTLE_KING_WALK 🔒

WHITE_LONG_CASTLE_ROOK_WALK 🔒

WHITE_SHORT_CASTLE_KING_WALK 🔒

WHITE_SHORT_CASTLE_ROOK_WALK 🔒

Functions

bishop_attacks 🔒

bishop_ray 🔒

king_attacks 🔒

knight_attacks 🔒

pawn_attacks 🔒

pext 🔒

queen_attacks 🔒

ray 🔒

rook_attacks 🔒

rook_ray 🔒