Struct pabi::chess::bitboard::Bitboard

pub struct Bitboard {
    bits: u64,
}

Represents a set of squares and provides common operations (e.g. AND, OR, XOR) over these sets. Each bit corresponds to one of 64 squares of the chess board.

Mirroring Square semantics, the least significant bit corresponds to Square::A1, and the most significant bit - to Square::H8.

Bitboard is a thin wrapper around u64:

use std::mem::size_of;

use pabi::chess::bitboard::Bitboard;

assert_eq!(size_of::<Bitboard>(), 8);

Fields
bits: u64
Implementations

impl Bitboard

pub const fn from_bits(bits: u64) -> Self

Constructs Bitboard from pre-calculated bits.

pub const fn empty() -> Self

Constructs a bitboard representing empty set of squares.

pub const fn full() -> Self

Constructs a bitboard representing the universal set, it contains all squares by setting all bits to binary one.

pub const fn bits(self) -> u64

Returns raw bits.

pub fn from_squares(squares: &[Square]) -> Self

pub(super) fn extend(&mut self, square: Square)

Adds given square to the set.

pub(super) fn clear(&mut self, square: Square)

Clears given square from the set.

pub(super) const fn contains(self, square: Square) -> bool

Returns true if this bitboard contains given square.

pub(super) const fn as_square(self) -> Square

pub(crate) const fn count(self) -> u32

pub(super) const fn is_empty(self) -> bool

pub(super) const fn has_any(self) -> bool

pub(super) fn shift(self, direction: Direction) -> Self

pub fn flip_perspective(&self) -> Self

Flips the bitboard vertically.

This is useful when we want to switch between the board point of view of White and Black.

Example

use pabi::chess::bitboard::Bitboard;

let bb = Bitboard::from_bits(0x1e2222120e0a1222);
assert_eq!(
    format!("{:?}", bb),
    ". 1 1 1 1 . . .\n\
     . 1 . . . 1 . .\n\
     . 1 . . . 1 . .\n\
     . 1 . . 1 . . .\n\
     . 1 1 1 . . . .\n\
     . 1 . 1 . . . .\n\
     . 1 . . 1 . . .\n\
     . 1 . . . 1 . ."
);
assert_eq!(
    format!("{:?}", bb.flip_perspective()),
    ". 1 . . . 1 . .\n\
     . 1 . . 1 . . .\n\
     . 1 . 1 . . . .\n\
     . 1 1 1 . . . .\n\
     . 1 . . 1 . . .\n\
     . 1 . . . 1 . .\n\
     . 1 . . . 1 . .\n\
     . 1 1 1 1 . . ."
);

pub(super) const fn iter(self) -> BitboardIterator

An efficient way to iterate over the set squares.